pypar.balance(): guard against out-of-bound bin index
In pypar.balance(), make it so that whenever the local id (p) is larger
that the index of the last bin, the returned slice bounds gives a
zero-length slice (N, N).

Before the change, this is not true if the number of bins (P) is not
greater than the upper bound of full interval (N).